Bug 1432678. Convert to size_t in DecodedSurfaceProvider::LogicalSizeInBytes so multiplication doesn't overflow. r=aosmond
UltraBlame original commit: dff22255b582c72ad2a46f09d10b4410029a3e2a

+++ image/DecodedSurfaceProvider.cpp
@@ -122,7 +122,7 @@ DecodedSurfaceProvider::LogicalSizeInBytes() const
 {
   
   IntSize size = GetSurfaceKey().Size();
-  return size.width * size.height * sizeof(uint32_t);
+  return size_t(size.width) * size_t(size.height) * sizeof(uint32_t);
 }
 
 void
